core: reviewed types
