+++ client/core/chain_makers.go
@@ -80,7 +80,7 @@ func (b *BlockGen) AddTxWithChain(bc *BlockChain, tx *types.Transaction) {
 		b.SetCoinbase(common.Address{})
 	}
 	b.statedb.Prepare(tx.Hash(), common.Hash{}, len(b.txs))
-	receipt, _, err := ApplyTransaction(b.config, bc, &b.header.Coinbase, b.crpool, b.statedb, b.header, tx, &b.header.ComputationalEffort, vm.Config{})
+	receipt, _, err := ApplyTransaction(b.config, bc, &b.header.Coinbase, b.crpool, b.statedb, b.header, tx, &b.header.ResourceUsage, vm.Config{})
 	if err != nil {
 		panic(err)
 	}
