+++ client/core/tx_pool.go
@@ -1,358 +1,161 @@
 package core
 
-import (
-	"errors"
-	"sync"
-	"time"
-
-	"github.com/kowala-tech/kcoin/client/common"
-	"github.com/kowala-tech/kcoin/client/core/state"
-	"github.com/kowala-tech/kcoin/client/core/types"
-	"github.com/kowala-tech/kcoin/client/event"
-	"github.com/kowala-tech/kcoin/client/log"
-	"github.com/kowala-tech/kcoin/client/metrics"
-	"github.com/kowala-tech/kcoin/client/params"
-)
-
-// TODO - add tx queue in tx pool init
-
-const (
-	// chainHeadChSize is the size of the channel listening to the ChainHeadEvent.
-	chainHeahChSize = 10
-)
-
-var (
-	// ErrInvalidSender is returned if the transaction contains an invalid signature
-	ErrInvalidSender = errors.New("invalid sender")
-
-	// ErrNonceTooLow is returned if the nonce of a transaction is lower than the
-	// one present in the local chain.
-	ErrNonceTooLow = errors.New("nonce too low")
-
-	// ErrInsufficientFunds is returned if the total cost of executing a transaction
-	// is higher than the balance of the user's account.
-	ErrInsufficientFunds = errors.New("insufficient funds for compute units * price + value")
-
-	// ErrIntrinsicResources is returned if the transaction is specified
-	// to use less compute resources than required to start the invocation.
-	ErrIntrinsicResources = errors.New("intrinsic compute resources low")
-
-	// ErrComputeCapacity is returned if a transaction's requested compute
-	// resources exceeds the compute capacity of the current block.
-	ErrComputeCapacity = errors.New("exceeds block compute capacity")
-
-	// ErrNegativeValue is a sanity error to ensure none is able to specify a
-	// transaction with a negative value.
-	ErrNegativeValue = errors.New("negative value")
-
-	// ErrOversizedData is returned if the input data of a transaction is greater
-	// than some meaningful limit a user might use. This is not a consensus error
-	// making the transaction invalid, rather a DOS protection.
-	ErrOversizedData = errors.New("oversized data")
-)
-
-var (
-	// Metrics for the executable pool
-	pendingDiscardCounter   = metrics.NewRegisteredCounter("txpool/executable/discard", nil)
-	pendingRateLimitCounter = metrics.NewRegisteredCounter("txpool/executable/ratelimit", nil) // Dropped due to rate limiting
-	pendingNofundsCounter   = metrics.NewRegisteredCounter("txpool/executable/nofunds", nil)   // Dropped due to out-of-funds
-
-	// Metrics for the pending pool
-	queuedDiscardCounter   = metrics.NewRegisteredCounter("txpool/pending/discard", nil)
-	queuedRateLimitCounter = metrics.NewRegisteredCounter("txpool/pending/ratelimit", nil) // Dropped due to rate limiting
-	queuedNofundsCounter   = metrics.NewRegisteredCounter("txpool/pending/nofunds", nil)   // Dropped due to out-of-funds
-
-	// General tx metrics
-	invalidTxCounter = metrics.NewRegisteredCounter("txpool/invalid", nil)
-)
-
-// TxStatus is the current status of a transaction as seen by the pool.
-type TxStatus uint
-
-const (
-	TxStatusUnknown TxStatus = iota
-	TxStatusQueued
-	TxStatusPending
-	TxStatusIncluded
-)
-
-// blockChain provides the state of blockchain and current compute capacity to do
-// some pre checks in tx pool and event subscribers.
-type blockChain interface {
-	CurrentBlock() *types.Block
-	GetBlock(hash common.Hash, number uint64) *types.Block
-	StateAt(root common.Hash) (*state.StateDB, error)
-
-	SubscribeChainHeadEvent(ch chan<- ChainHeadEvent) event.Subscription
-}
-
-// DefaultTxPoolConfig contains the default configurations for the transaction
-// pool.
-var DefaultTxPoolConfig = TxPoolConfig{
-	Journal:   "txjournal.kowala",
-	Rejournal: time.Hour,
-
-	AccountSlots: 16,
-	GlobalSlots:  4096,
-	AccountQueue: 64,
-	GlobalQueue:  1024,
-
-	Lifetime: 3 * time.Hour,
-
-	StatsReportInterval: 8 * time.Second,
-}
-
-// TxPoolConfig are the configurable parameters of the transaction pool.
-type TxPoolConfig struct {
-	NoLocals  bool          // Whether local transaction handling should be disabled
-	Journal   string        // Journal of local transactions to survive node restarts
-	Rejournal time.Duration // Time interval to regenerate the local transaction journal
-
-	AccountSlots uint64 // Minimum number of executable transaction slots guaranteed per account
-	GlobalSlots  uint64 // Maximum number of executable transaction slots for all accounts
-	AccountQueue uint64 // Maximum number of non-executable transaction slots permitted per account
-	GlobalQueue  uint64 // Maximum number of non-executable transaction slots for all accounts
-
-	Lifetime time.Duration // Maximum amount of time non-executable transaction are queued
-
-	StatsReportInterval time.Duration // Time interval to report transaction pool stats
-}
-
-// sanitize checks the provided user configurations and changes anything that's
-// unreasonable or unworkable.
-func (config *TxPoolConfig) sanitize() TxPoolConfig {
-	conf := *config
-	if conf.Rejournal < time.Second {
-		log.Warn("Sanitizing invalid txpool journal time", "provided", conf.Rejournal, "updated", time.Second)
-		conf.Rejournal = time.Second
-	}
-	return conf
-}
-
-// TxPool contains all currently known transactions. Transactions
-// enter the pool when they are received from the network or submitted
-// locally. They exit the pool when they are included in the blockchain.
-//
-// The pool separates pending transactions from executable transactions
-// (which can be applied to the current state). Transactions move between those
-// two states over time as they are received and processed.
 type TxPool struct {
-	mu           sync.RWMutex
-	config       TxPoolConfig
-	chainconfig  *params.ChainConfig
-	chain        blockChain
-	txFeed       event.Feed
-	scope        event.SubscriptionScope
-	chainHeadCh  chan ChainHeadEvent
-	chainHeadSub event.Subscription
-
-	signer types.Signer
-
-	currentState *state.StateDB      // Current state in the blockchain head
-	pendingState *state.ManagedState // Pending state tracking virtual nonces
-
-	locals  *accountSet // Set of local transactions
-	journal *txJournal  // Journal of local transactions to back up to disk
-
-	ready   *txQueue                   // All currently processable transactions
-	pending map[common.Address]*txList // Queued but non-processable transactions
-	all     *txLookup                  // All transactions to allow lookups
-
-	wg sync.WaitGroup // for shutdown sync
+	byNonce map[common.Address]*txList
 }
 
-// NewTxPool creates a new transaction pool to gather, sort and filter
-// inbound transactions from the network
-func NewTxPool(config TxPoolConfig, chainconfig *params.ChainConfig, chain blockChain) *TxPool {
-	config = (&config).sanitize()
-
-	pool := &TxPool{
-		config:      config,
-		chainconfig: chainconfig,
-		chain:       chain,
-		signer:      types.NewAndromedaSigner(chainconfig.ChainID),
-		ready:       newTxQueue(),
-		pending:     make(map[common.Address]*txList),
-		all:         newTxLookup(),
-		chainHeadCh: make(chan ChainHeadEvent, chainHeadChSize),
-	}
-
-	pool.locals = newAccountSet(pool.signer)
-	pool.reset(nil, chain.CurrentBlock().Header())
-
-	// If local transactions and journaling is enabled, load from disk
-	if !config.NoLocals && config.Journal != "" {
-		pool.journal = newTxJournal(config.Journal)
-
-		if err := pool.journal.load(pool.AddLocals); err != nil {
-			log.Warn("Failed to load transaction journal", "err", err)
-		}
-		if err := pool.journal.rotate(pool.local()); err != nil {
-			log.Warn("Failed to rotate transaction journal", "err", err)
-		}
-	}
-
-	pool.chainHeadSub = pool.chain.SubscribeChainHeadEvent(pool.chainHeadCh)
-
-	pool.wg.Add(1)
-	go pool.loop()
-
-	return pool
+// AddLocal enqueues a single transaction into the pool if it is valid, making the sender 
+// as local if the distinction between local/remote transactions is enabled.
+func (pool *TxPool) AddLocal(tx *types.Transaction) error {
+	return pool.addTx(tx, !pool.config.NoLocals)
 }
 
-// loop is the transaction pool's main event loop, waiting for and reacting to
-// outside blockchain events as well as for various reporting events.
-func (pool *TxPool) loop() {
-	defer pool.wg.Done()
+// AddRemote enqueues a single remote transaction into the pool if it is valid.
+func (pool *TxPool) AddRemote(tx *types.Transaction) error {
+	return pool.addTx(tx, false)
+}
 
-	var prevReady, prevPending int
+// AddLocals enqueues a batch of transactions into the pool if they are valid,
+// marking the senders as a local ones if the distinction between local/remote 
+// transactions is enabled.
+func (pool *TxPool) AddLocals(txs []*types.Transaction) []error {
+	return pool.addTxs(txs, !pool.config.NoLocals)
+}
 
-	// start the stats reporting/journal tickers
-	report := time.NewTicker(statsReportInterval)
-	defer report.Stop()
-	journal := time.NewTicker(pool.config.Rejournal)
-	defer journal.Stop()
+// AddRemotes enqueues a batch of remote transactions into the pool if they are valid.
+func (pool *TxPool) AddRemotes(txs []*types.Transaction) []error {
+	return pool.addTxs(txs, false)
+}
 
-	// track the previous head headers for transaction reorgs
-	head := pool.chain.CurrentBlock()
+// addTxs attempts to queue a batch of transactions if they are valid.
+func (pool *TxPool) addTxs(txs []*types.Transaction, local bool) []error {
+	pool.mu.Lock()
+	defer pool.mu.Unlock()
 
-	// keep waiting for and reacting to the various events
-	for {
-		select {
-		// Handle ChainHeadEvent
-		case ev := <-pool.chainHeadCh:
-			if ev.Block != nil {
-				pool.mu.Lock()
-				pool.reset(head.Header(), ev.Block.Header())
-				head = ev.Block
+	return pool.addTxsLocked(txs, local)
+}
 
-				pool.mu.Unlock()
-			}
-		// be unsubscribed due to system stopped
-		case <-pool.chainHeadSub.Err():
-			return
+// addTx enqueues a single transaction into the pool if it is valid.
+func (pool *TxPool) addTx(tx *types.Transaction, local bool) error {
+	pool.mu.Lock()
+	defer pool.mu.Unlock()
 
-		// handle stats reporting ticks
-		case <-report.C:
-			pool.mu.RLock()
-			ready, pending := pool.stats()
-			pool.mu.RUnlock()
+	// Try to inject the transaction and update any state
+	replace, err := pool.add(tx, local)
+	if err != nil {
+		return err
+	}
+	// If we added a new transaction, run promotion checks and return
+	if !replace {
+		from, _ := types.TxSender(pool.signer, tx) // already validated
+		pool.promoteExecutables([]common.Address{from})
+	}
+	return nil
+}
 
-			if ready != prevReady || pending != prevPending {
-				log.Debug("Transaction pool status report", "ready", ready, "pending", pending)
-				prevReady, prevPending = ready, pending
-			}
+// addTxsLocked attempts to queue a batch of transactions if they are valid,
+// whilst assuming the transaction pool lock is already held.
+func (pool *TxPool) addTxsLocked(txs []*types.Transaction, local bool) []error {
+	// Add the batch of transaction, tracking the accepted ones
+	dirty := make(map[common.Address]struct{})
+	errs := make([]error, len(txs))
 
-		// handle local transaction journal rotation
-		case <-journal.C:
-			if pool.journal != nil {
-				pool.mu.Lock()
-				if err := pool.journal.rotate(pool.local()); err != nil {
-					log.Warn("Failed to rotate local tx journal", "err", err)
-				}
-				pool.mu.Unlock()
-			}
+	for i, tx := range txs {
+		var replace bool
+		if replace, errs[i] = pool.add(tx, local); errs[i] == nil && !replace {
+			from, _ := types.TxSender(pool.signer, tx) // already validated
+			dirty[from] = struct{}{}
 		}
 	}
-}
-
-// txLookup is used internally by TxPool to track transactions while allowing lookup without
-// mutex contention.
-//
-// Note, although this type is properly protected against concurrent access, it
-// is **not** a type that should ever be mutated or even exposed outside of the
-// transaction pool, since its internal state is tightly coupled with the pools
-// internal mechanisms. The sole purpose of the type is to permit out-of-bound
-// peeking into the pool in TxPool.Get without having to acquire the widely scoped
-// TxPool.mu mutex.
-type txLookup struct {
-	all  map[common.Hash]*types.Transaction
-	lock sync.RWMutex
-}
-
-// newTxLookup returns a new txLookup structure.
-func newTxLookup() *txLookup {
-	return &txLookup{
-		all: make(map[common.Hash]*types.Transaction),
-	}
-}
-
-// Range calls f on each key and value present in the map.
-func (t *txLookup) Range(f func(hash common.Hash, tx *types.Transaction) bool) {
-	t.lock.RLock()
-	defer t.lock.RUnlock()
-
-	for key, value := range t.all {
-		if !f(key, value) {
-			break
+	// Only reprocess the internal state if something was actually added
+	if len(dirty) > 0 {
+		addrs := make([]common.Address, 0, len(dirty))
+		for addr := range dirty {
+			addrs = append(addrs, addr)
 		}
+		pool.promoteExecutables(addrs)
 	}
+	return errs
 }
 
-// Get returns a transaction if it exists in the lookup, or nil if not found.
-func (t *txLookup) Get(hash common.Hash) *types.Transaction {
-	t.lock.RLock()
-	defer t.lock.RUnlock()
-
-	return t.all[hash]
-}
-
-// Count returns the current number of items in the lookup.
-func (t *txLookup) Count() int {
-	t.lock.RLock()
-	defer t.lock.RUnlock()
-
-	return len(t.all)
-}
-
-// Add adds a transaction to the lookup.
-func (t *txLookup) Add(tx *types.Transaction) {
-	t.lock.Lock()
-	defer t.lock.Unlock()
-
-	t.all[tx.Hash()] = tx
-}
-
-// Remove removes a transaction from the lookup.
-func (t *txLookup) Remove(hash common.Hash) {
-	t.lock.Lock()
-	defer t.lock.Unlock()
-
-	delete(t.all, hash)
-}
-
-// accountSet is simply a set of addresses to check for existence, and a signer
-// capable of deriving addresses from transactions.
-type accountSet struct {
-	accounts map[common.Address]struct{}
-	signer   types.Signer
-}
-
-// newAccountSet creates a new address set with an associated signer for sender
-// derivations.
-func newAccountSet(signer types.Signer) *accountSet {
-	return &accountSet{
-		accounts: make(map[common.Address]struct{}),
-		signer:   signer,
+// add validates a transaction and inserts it into the pending queue for
+// later promotion and execution.
+func (pool *TxPool) add(tx *types.Transaction, local bool) (bool, error) {
+	// If the transaction is already known, discard it
+	hash := tx.Hash()
+	if pool.all.Get(hash) != nil {
+		log.Trace("Discarding already known transaction", "hash", hash)
+		return false, fmt.Errorf("known transaction: %x", hash)
 	}
-}
-
-// contains checks if a given address is contained within the set.
-func (as *accountSet) contains(addr common.Address) bool {
-	_, exist := as.accounts[addr]
-	return exist
-}
-
-// containsTx checks if the sender of a given tx is within the set. If the sender
-// cannot be derived, this method returns false.
-func (as *accountSet) containsTx(tx *types.Transaction) bool {
-	if addr, err := types.TxSender(as.signer, tx); err == nil {
-		return as.contains(addr)
+	// If the transaction fails basic validation, discard it
+	if err := pool.validateTx(tx, local); err != nil {
+		log.Trace("Discarding invalid transaction", "hash", hash, "err", err)
+		invalidTxCounter.Inc(1)
+		return false, err
 	}
-	return false
+
+	// If the transaction overlaps an existing one (repeated nonce)
+	from, _ := types.TxSender(pool.signer, tx)
+	if list := pool.byNonce[from]; list != nil && list.Overlaps(tx) {
+		log.Trace("Discarding transaction that overlaps an existing one", "nonce", tx.Nonce(), "err", err)
+		invalidTxCounter.Inc(1)
+		return false, fmt.Errorf("known transaction: %x", hash)
+	}
+
+	replace, err := pool.enqueueTx(hash, tx)
+	if err != nil {
+		return false, err
+	}
+	// Mark local addresses and journal local transactions
+	if local {
+		pool.locals.add(from)
+	}
+	pool.journalTx(from, tx)
+
+	log.Trace("Pooled new future transaction", "hash", hash, "from", from, "to", tx.To())
+	return replace, nil
 }
 
-// add inserts a new address into the set to track.
-func (as *accountSet) add(addr common.Address) {
-	as.accounts[addr] = struct{}{}
+// validateTx checks whether a transaction is valid according to the consensus
+// rules and adheres to some heuristic limits (ex: size).
+func (pool *TxPool) validateTx(tx *types.Transaction, local bool) error {
+	// Heuristic limit, reject transactions over the maximum size to prevent DOS attacks
+	if tx.Size() > params.MaxTxSize {
+		return ErrOversizedData
+	}
+	// Transactions can't be negative. This may never happen using RLP decoded
+	// transactions but may occur if you create a transaction using the RPC.
+	if tx.Value().Sign() < 0 {
+		return ErrNegativeValue
+	}
+	// Ensure the transaction doesn't exceed the block compute capacity.
+	if params.ComputeCapacity < tx.ComputeLimit() {
+		return ErrComputeCapacity
+	}
+	// Make sure the transaction is signed properly
+	from, err := types.TxSender(pool.signer, tx)
+	if err != nil {
+		return ErrInvalidSender
+	}
+
+	// Ensure the transaction has a valid nonce
+	if pool.currentState.GetNonce(from) > tx.Nonce() {
+		return ErrNonceTooLow
+	}
+
+	// Sender should have enough funds to cover the costs
+	// cost == value + computational effort (compute units) * compute unit price
+	if pool.currentState.GetBalance(from).Cmp(tx.Cost()) < 0 {
+		return ErrInsufficientFunds
+	}
+
+	intrCompEffort, err := IntrinsicCompEffort(tx.Data(), tx.To() == nil, true)
+	if err != nil {
+		return err
+	}
+	if tx.ComputeLimit() < intrCompEffort {
+		return ErrIntrinsicCompEffort
+	}
+	return nil
 }
+
