+++ client/core/types/gen_header_json.go
@@ -25,6 +25,7 @@ func (h Header) MarshalJSON() ([]byte, error) {
 		LastCommitHash common.Hash    `json:"lastCommit"       gencodec:"required"`
 		Bloom          Bloom          `json:"logsBloom"        gencodec:"required"`
 		Number         *hexutil.Big   `json:"number"           gencodec:"required"`
+		ResourceUsage  hexutil.Uint64 `json:"resourceUsage"    gencodec:"required"`
 		Time           *hexutil.Big   `json:"timestamp"        gencodec:"required"`
 		Extra          hexutil.Bytes  `json:"extraData"        gencodec:"required"`
 		Hash           common.Hash    `json:"hash"`
@@ -39,6 +40,7 @@ func (h Header) MarshalJSON() ([]byte, error) {
 	enc.LastCommitHash = h.LastCommitHash
 	enc.Bloom = h.Bloom
 	enc.Number = (*hexutil.Big)(h.Number)
+	enc.ResourceUsage = hexutil.Uint64(h.ResourceUsage)
 	enc.Time = (*hexutil.Big)(h.Time)
 	enc.Extra = h.Extra
 	enc.Hash = h.Hash()
@@ -57,6 +59,7 @@ func (h *Header) UnmarshalJSON(input []byte) error {
 		LastCommitHash *common.Hash    `json:"lastCommit"       gencodec:"required"`
 		Bloom          *Bloom          `json:"logsBloom"        gencodec:"required"`
 		Number         *hexutil.Big    `json:"number"           gencodec:"required"`
+		ResourceUsage  *hexutil.Uint64 `json:"resourceUsage"    gencodec:"required"`
 		Time           *hexutil.Big    `json:"timestamp"        gencodec:"required"`
 		Extra          *hexutil.Bytes  `json:"extraData"        gencodec:"required"`
 	}
@@ -100,6 +103,10 @@ func (h *Header) UnmarshalJSON(input []byte) error {
 		return errors.New("missing required field 'number' for Header")
 	}
 	h.Number = (*big.Int)(dec.Number)
+	if dec.ResourceUsage == nil {
+		return errors.New("missing required field 'resourceUsage' for Header")
+	}
+	h.ResourceUsage = uint64(*dec.ResourceUsage)
 	if dec.Time == nil {
 		return errors.New("missing required field 'timestamp' for Header")
 	}
